@@ -1,13 +1,13 @@
 name: connectivity_plus
 description: Flutter plugin for discovering the state of the network (WiFi & mobile/cellular) connectivity on Android and iOS.
-version: 3.0.2
+version: 4.0.1
 homepage: https://plus.fluttercommunity.dev/
-repository: https://github.com/fluttercommunity/plus_plugins/tree/main/packages/
+repository: https://github.com/fluttercommunity/plus_plugins/tree/main/packages/connectivity_plus/connectivity_plus
 issue_tracker: https://github.com/fluttercommunity/plus_plugins/labels/connectivity_plus
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=2.11.0"
+  sdk: ">=2.18.0 <4.0.0"
+  flutter: ">=3.3.0"
 
 flutter:
   plugin:
@@ -32,17 +32,17 @@ dependencies:
     sdk: flutter
   flutter_web_plugins:
     sdk: flutter
-  connectivity_plus_platform_interface: ^1.2.3
-  js: ^0.6.3
+  connectivity_plus_platform_interface: ^1.2.4
+  js: ^0.6.5
   meta: ^1.8.0
   nm: ^0.5.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: ^2.1.11
-  dbus: ^0.7.5
+  build_runner: ^2.3.3
+  dbus: ^0.7.8
   flutter_lints: ^2.0.1
-  mockito: ^5.2.0
-  plugin_platform_interface: ^2.1.2
-  test: ^1.21.1
+  mockito: ^5.4.0
+  plugin_platform_interface: ^2.1.4
+  test: ^1.22.0
