@@ -1,9 +1,6 @@
-//@dart=2.9
-
 // Copyright 2019 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
-
 import 'dart:io';
 
 import 'package:connectivity_plus/connectivity_plus.dart';
@@ -13,9 +10,9 @@ import 'package:integration_test/integration_test.dart';
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
-  group('Connectivity test driver', () {
-    Connectivity connectivity;
+  late Connectivity connectivity;
 
+  group('Connectivity test driver', () {
     setUpAll(() async {
       connectivity = Connectivity();
     });
@@ -25,12 +22,23 @@ void main() {
       expect(result, isNotNull);
     });
 
-    testWidgets('connectivity on Android emulator should be wifi',
+    testWidgets('connectivity on Android newer than 5 (API 21) should be wifi',
         (WidgetTester tester) async {
       final result = await connectivity.checkConnectivity();
 
       expect(result, ConnectivityResult.wifi);
-    }, skip: !Platform.isAndroid);
+    },
+        skip: !Platform.isAndroid ||
+            Platform.operatingSystemVersion.contains('5.0.2'));
+
+    testWidgets('connectivity on Android 5 (API 21) should be mobile',
+        (WidgetTester tester) async {
+      final result = await connectivity.checkConnectivity();
+
+      expect(result, ConnectivityResult.mobile);
+    },
+        skip: !Platform.isAndroid ||
+            !Platform.operatingSystemVersion.contains('5.0.2'));
 
     testWidgets('connectivity on MacOS should be ethernet',
         (WidgetTester tester) async {
