@@ -35,7 +35,8 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       title: 'Flutter Demo',
       theme: ThemeData(
-        primarySwatch: Colors.blue,
+        useMaterial3: true,
+        colorSchemeSeed: const Color(0x9f4376f8),
       ),
       home: const MyHomePage(title: 'Flutter Demo Home Page'),
     );
@@ -66,6 +67,7 @@ class _MyHomePageState extends State<MyHomePage> {
     return Scaffold(
       appBar: AppBar(
         title: const Text('NetworkInfoPlus example'),
+        elevation: 4,
       ),
       body: Center(
           child: Column(
@@ -96,8 +98,10 @@ class _MyHomePageState extends State<MyHomePage> {
 
     try {
       if (!kIsWeb && Platform.isIOS) {
+        // ignore: deprecated_member_use
         var status = await _networkInfo.getLocationServiceAuthorization();
         if (status == LocationAuthorizationStatus.notDetermined) {
+          // ignore: deprecated_member_use
           status = await _networkInfo.requestLocationServiceAuthorization();
         }
         if (status == LocationAuthorizationStatus.authorizedAlways ||
@@ -116,8 +120,10 @@ class _MyHomePageState extends State<MyHomePage> {
 
     try {
       if (!kIsWeb && Platform.isIOS) {
+        // ignore: deprecated_member_use
         var status = await _networkInfo.getLocationServiceAuthorization();
         if (status == LocationAuthorizationStatus.notDetermined) {
+          // ignore: deprecated_member_use
           status = await _networkInfo.requestLocationServiceAuthorization();
         }
         if (status == LocationAuthorizationStatus.authorizedAlways ||
