@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:device_info_plus/device_info_plus.dart';
 import 'package:device_info_plus_platform_interface/device_info_plus_platform_interface.dart';
 import 'package:file/file.dart';
 import 'package:file/local.dart';
@@ -20,8 +21,12 @@ class DeviceInfoPlusLinuxPlugin extends DeviceInfoPlatform {
       : _fileSystem = fileSystem ?? const LocalFileSystem();
 
   @override
-  Future<LinuxDeviceInfo> linuxInfo() async {
+  Future<BaseDeviceInfo> deviceInfo() async {
     return _cache ??= await _getInfo();
+  }
+
+  Future<LinuxDeviceInfo> linuxInfo() async {
+    return (await deviceInfo()) as LinuxDeviceInfo;
   }
 
   Future<LinuxDeviceInfo> _getInfo() async {
