@@ -4,6 +4,8 @@ import 'dart:html' as html show window, Navigator;
 import 'package:device_info_plus_platform_interface/device_info_plus_platform_interface.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 
+import 'model/web_browser_info.dart';
+
 /// The web implementation of the BatteryPlusPlatform of the BatteryPlus plugin.
 class DeviceInfoPlusWebPlugin extends DeviceInfoPlatform {
   /// Constructs a DeviceInfoPlusPlugin.
@@ -19,7 +21,7 @@ class DeviceInfoPlusWebPlugin extends DeviceInfoPlatform {
   }
 
   @override
-  Future<WebBrowserInfo> webBrowserInfo() {
+  Future<BaseDeviceInfo> deviceInfo() {
     return Future<WebBrowserInfo>.value(
       WebBrowserInfo.fromMap(
         {
