@@ -1,4 +1,7 @@
-import 'package:device_info_plus_platform_interface/device_info_plus_platform_interface.dart';
+@TestOn('browser')
+library device_info_plus_web_test;
+
+import 'package:device_info_plus/src/model/web_browser_info.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
