@@ -1,6 +1,17 @@
 /// The base class for platform's device info.
-abstract class BaseDeviceInfo {
-  /// Serializes device info properties to a map.
-  @Deprecated('[toMap] method will be discontinued')
-  Map<String, dynamic> toMap();
+class BaseDeviceInfo {
+  BaseDeviceInfo(this.data);
+
+  /// Device information data
+  /// Warning: The returned Map may not be JSON-encodable.
+  final Map<String, dynamic> data;
+
+  /// For legacy purposes
+  @Deprecated('Use [data] getter instead')
+  Map<String, dynamic> toMap() => data;
+
+  @override
+  String toString() {
+    return 'BaseDeviceInfo{data: $data}';
+  }
 }
