@@ -6,9 +6,13 @@
 
 import 'dart:io';
 
-import 'package:file_selector/file_selector.dart';
+import 'package:file_selector/file_selector.dart'
+    hide XFile; // hides to test if share_plus exports XFile
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:image_picker/image_picker.dart';
+import 'package:flutter/services.dart';
+import 'package:image_picker/image_picker.dart'
+    hide XFile; // hides to test if share_plus exports XFile
 import 'package:share_plus/share_plus.dart';
 
 import 'image_previews.dart';
@@ -27,6 +31,7 @@ class DemoApp extends StatefulWidget {
 class DemoAppState extends State<DemoApp> {
   String text = '';
   String subject = '';
+  String uri = '';
   List<String> imageNames = [];
   List<String> imagePaths = [];
 
@@ -34,99 +39,143 @@ class DemoAppState extends State<DemoApp> {
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Share Plus Plugin Demo',
+      theme: ThemeData(
+        useMaterial3: true,
+        colorSchemeSeed: const Color(0x9f4376f8),
+      ),
       home: Scaffold(
-          appBar: AppBar(
-            title: const Text('Share Plus Plugin Demo'),
-          ),
-          body: SingleChildScrollView(
-            child: Padding(
-              padding: const EdgeInsets.all(24.0),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: <Widget>[
-                  TextField(
-                    decoration: const InputDecoration(
-                      labelText: 'Share text:',
-                      hintText: 'Enter some text and/or link to share',
-                    ),
-                    maxLines: 2,
-                    onChanged: (String value) => setState(() {
-                      text = value;
-                    }),
-                  ),
-                  TextField(
-                    decoration: const InputDecoration(
-                      labelText: 'Share subject:',
-                      hintText: 'Enter subject to share (optional)',
-                    ),
-                    maxLines: 2,
-                    onChanged: (String value) => setState(() {
-                      subject = value;
-                    }),
-                  ),
-                  const Padding(padding: EdgeInsets.only(top: 12.0)),
-                  ImagePreviews(imagePaths, onDelete: _onDeleteImage),
-                  ListTile(
-                    leading: const Icon(Icons.add),
-                    title: const Text('Add image'),
-                    onTap: () async {
-                      // Using `package:image_picker` to get image from gallery.
-                      if (Platform.isMacOS ||
-                          Platform.isLinux ||
-                          Platform.isWindows) {
-                        // Using `package:file_selector` on windows, macos & Linux, since `package:image_picker` is not supported.
-                        const XTypeGroup typeGroup = XTypeGroup(
-                          label: 'images',
-                          extensions: <String>['jpg', 'jpeg', 'png', 'gif'],
-                        );
-                        final file = await openFile(
-                            acceptedTypeGroups: <XTypeGroup>[typeGroup]);
-                        if (file != null) {
-                          setState(() {
-                            imagePaths.add(file.path);
-                            imageNames.add(file.name);
-                          });
-                        }
-                      } else {
-                        final imagePicker = ImagePicker();
-                        final pickedFile = await imagePicker.pickImage(
-                          source: ImageSource.gallery,
-                        );
-                        if (pickedFile != null) {
-                          setState(() {
-                            imagePaths.add(pickedFile.path);
-                            imageNames.add(pickedFile.name);
-                          });
-                        }
-                      }
-                    },
-                  ),
-                  const Padding(padding: EdgeInsets.only(top: 12.0)),
-                  Builder(
-                    builder: (BuildContext context) {
-                      return ElevatedButton(
-                        onPressed: text.isEmpty && imagePaths.isEmpty
-                            ? null
-                            : () => _onShare(context),
-                        child: const Text('Share'),
-                      );
-                    },
-                  ),
-                  const Padding(padding: EdgeInsets.only(top: 12.0)),
-                  Builder(
-                    builder: (BuildContext context) {
-                      return ElevatedButton(
-                        onPressed: text.isEmpty && imagePaths.isEmpty
-                            ? null
-                            : () => _onShareWithResult(context),
-                        child: const Text('Share With Result'),
-                      );
-                    },
-                  ),
-                ],
+        appBar: AppBar(
+          title: const Text('Share Plus Plugin Demo'),
+          elevation: 4,
+        ),
+        body: SingleChildScrollView(
+          padding: const EdgeInsets.all(24),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            mainAxisSize: MainAxisSize.min,
+            children: <Widget>[
+              TextField(
+                decoration: const InputDecoration(
+                  border: OutlineInputBorder(),
+                  labelText: 'Share text',
+                  hintText: 'Enter some text and/or link to share',
+                ),
+                maxLines: null,
+                onChanged: (String value) => setState(() {
+                  text = value;
+                }),
               ),
-            ),
-          )),
+              const SizedBox(height: 16),
+              TextField(
+                decoration: const InputDecoration(
+                  border: OutlineInputBorder(),
+                  labelText: 'Share subject',
+                  hintText: 'Enter subject to share (optional)',
+                ),
+                maxLines: null,
+                onChanged: (String value) => setState(() {
+                  subject = value;
+                }),
+              ),
+              const SizedBox(height: 16),
+              TextField(
+                decoration: const InputDecoration(
+                  border: OutlineInputBorder(),
+                  labelText: 'Share uri',
+                  hintText: 'Enter the uri you want to share',
+                ),
+                maxLines: null,
+                onChanged: (String value) {
+                  setState(() => uri = value);
+                },
+              ),
+              const SizedBox(height: 16),
+              ImagePreviews(imagePaths, onDelete: _onDeleteImage),
+              ElevatedButton.icon(
+                label: const Text('Add image'),
+                onPressed: () async {
+                  // Using `package:image_picker` to get image from gallery.
+                  if (!kIsWeb &&
+                      (Platform.isMacOS ||
+                          Platform.isLinux ||
+                          Platform.isWindows)) {
+                    // Using `package:file_selector` on windows, macos & Linux, since `package:image_picker` is not supported.
+                    const XTypeGroup typeGroup = XTypeGroup(
+                      label: 'images',
+                      extensions: <String>['jpg', 'jpeg', 'png', 'gif'],
+                    );
+                    final file = await openFile(
+                        acceptedTypeGroups: <XTypeGroup>[typeGroup]);
+                    if (file != null) {
+                      setState(() {
+                        imagePaths.add(file.path);
+                        imageNames.add(file.name);
+                      });
+                    }
+                  } else {
+                    final imagePicker = ImagePicker();
+                    final pickedFile = await imagePicker.pickImage(
+                      source: ImageSource.gallery,
+                    );
+                    if (pickedFile != null) {
+                      setState(() {
+                        imagePaths.add(pickedFile.path);
+                        imageNames.add(pickedFile.name);
+                      });
+                    }
+                  }
+                },
+                icon: const Icon(Icons.add),
+              ),
+              const SizedBox(height: 32),
+              Builder(
+                builder: (BuildContext context) {
+                  return ElevatedButton(
+                    style: ElevatedButton.styleFrom(
+                      foregroundColor: Theme.of(context).colorScheme.onPrimary,
+                      backgroundColor: Theme.of(context).colorScheme.primary,
+                    ),
+                    onPressed: text.isEmpty && imagePaths.isEmpty && uri.isEmpty
+                        ? null
+                        : () => _onShare(context),
+                    child: const Text('Share'),
+                  );
+                },
+              ),
+              const SizedBox(height: 16),
+              Builder(
+                builder: (BuildContext context) {
+                  return ElevatedButton(
+                    style: ElevatedButton.styleFrom(
+                      foregroundColor: Theme.of(context).colorScheme.onPrimary,
+                      backgroundColor: Theme.of(context).colorScheme.primary,
+                    ),
+                    onPressed: text.isEmpty && imagePaths.isEmpty
+                        ? null
+                        : () => _onShareWithResult(context),
+                    child: const Text('Share With Result'),
+                  );
+                },
+              ),
+              const SizedBox(height: 16),
+              Builder(
+                builder: (BuildContext context) {
+                  return ElevatedButton(
+                    style: ElevatedButton.styleFrom(
+                      foregroundColor: Theme.of(context).colorScheme.onPrimary,
+                      backgroundColor: Theme.of(context).colorScheme.primary,
+                    ),
+                    onPressed: () {
+                      _onShareXFileFromAssets(context);
+                    },
+                    child: const Text('Share XFile from Assets'),
+                  );
+                },
+              ),
+            ],
+          ),
+        ),
+      ),
     );
   }
 
@@ -147,7 +196,9 @@ class DemoAppState extends State<DemoApp> {
     // has its position and size after it's built.
     final box = context.findRenderObject() as RenderBox?;
 
-    if (imagePaths.isNotEmpty) {
+    if (uri.isNotEmpty) {
+      await Share.shareUri(Uri.parse(uri));
+    } else if (imagePaths.isNotEmpty) {
       final files = <XFile>[];
       for (var i = 0; i < imagePaths.length; i++) {
         files.add(XFile(imagePaths[i], name: imageNames[i]));
@@ -166,23 +217,54 @@ class DemoAppState extends State<DemoApp> {
   void _onShareWithResult(BuildContext context) async {
     final box = context.findRenderObject() as RenderBox?;
     final scaffoldMessenger = ScaffoldMessenger.of(context);
-    ShareResult result;
+    ShareResult shareResult;
     if (imagePaths.isNotEmpty) {
       final files = <XFile>[];
       for (var i = 0; i < imagePaths.length; i++) {
         files.add(XFile(imagePaths[i], name: imageNames[i]));
       }
-      result = await Share.shareXFiles(files,
+      shareResult = await Share.shareXFiles(files,
           text: text,
           subject: subject,
           sharePositionOrigin: box!.localToGlobal(Offset.zero) & box.size);
     } else {
-      result = await Share.shareWithResult(text,
+      shareResult = await Share.shareWithResult(text,
           subject: subject,
           sharePositionOrigin: box!.localToGlobal(Offset.zero) & box.size);
     }
-    scaffoldMessenger.showSnackBar(SnackBar(
-      content: Text("Share result: ${result.status}"),
-    ));
+    scaffoldMessenger.showSnackBar(getResultSnackBar(shareResult));
+  }
+
+  void _onShareXFileFromAssets(BuildContext context) async {
+    final box = context.findRenderObject() as RenderBox?;
+    final scaffoldMessenger = ScaffoldMessenger.of(context);
+    final data = await rootBundle.load('assets/flutter_logo.png');
+    final buffer = data.buffer;
+    final shareResult = await Share.shareXFiles(
+      [
+        XFile.fromData(
+          buffer.asUint8List(data.offsetInBytes, data.lengthInBytes),
+          name: 'flutter_logo.png',
+          mimeType: 'image/png',
+        ),
+      ],
+      sharePositionOrigin: box!.localToGlobal(Offset.zero) & box.size,
+    );
+
+    scaffoldMessenger.showSnackBar(getResultSnackBar(shareResult));
+  }
+
+  SnackBar getResultSnackBar(ShareResult result) {
+    return SnackBar(
+      content: Column(
+        mainAxisSize: MainAxisSize.min,
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Text("Share result: ${result.status}"),
+          if (result.status == ShareResultStatus.success)
+            Text("Shared to: ${result.raw}")
+        ],
+      ),
+    );
   }
 }
