@@ -3,8 +3,8 @@ import 'package:share_plus/src/share_plus_windows.dart';
 import 'package:share_plus/src/windows_version_helper.dart';
 import 'package:share_plus_platform_interface/method_channel/method_channel_share.dart';
 import 'package:share_plus_platform_interface/share_plus_platform_interface.dart';
-import 'package:url_launcher_platform_interface/link.dart';
-import 'package:url_launcher_platform_interface/url_launcher_platform_interface.dart';
+
+import 'url_launcher_mock.dart';
 
 void main() {
   test(
@@ -31,9 +31,8 @@ void main() {
     'url encoding is correct for &',
     () async {
       final mock = MockUrlLauncherPlatform();
-      UrlLauncherPlatform.instance = mock;
 
-      await SharePlusWindowsPlugin().share('foo&bar', subject: 'bar&foo');
+      await SharePlusWindowsPlugin(mock).share('foo&bar', subject: 'bar&foo');
 
       expect(mock.url, 'mailto:?subject=bar%26foo&body=foo%26bar');
     },
@@ -45,9 +44,8 @@ void main() {
     'url encoding is correct for spaces',
     () async {
       final mock = MockUrlLauncherPlatform();
-      UrlLauncherPlatform.instance = mock;
 
-      await SharePlusWindowsPlugin().share('foo bar', subject: 'bar foo');
+      await SharePlusWindowsPlugin(mock).share('foo bar', subject: 'bar foo');
 
       expect(mock.url, 'mailto:?subject=bar%20foo&body=foo%20bar');
     },
@@ -59,39 +57,10 @@ void main() {
     () async {
       final mock = MockUrlLauncherPlatform();
       mock.canLaunchMockValue = false;
-      UrlLauncherPlatform.instance = mock;
 
-      expect(() async => await SharePlusWindowsPlugin().share('foo bar'),
+      expect(() async => await SharePlusWindowsPlugin(mock).share('foo bar'),
           throwsException);
     },
     skip: VersionHelper.instance.isWindows10RS5OrGreater,
   );
-}
-
-class MockUrlLauncherPlatform extends UrlLauncherPlatform {
-  String? url;
-  bool canLaunchMockValue = true;
-
-  @override
-  LinkDelegate? get linkDelegate => throw UnimplementedError();
-
-  @override
-  Future<bool> canLaunch(String url) async {
-    return canLaunchMockValue;
-  }
-
-  @override
-  Future<bool> launch(
-    String url, {
-    required bool useSafariVC,
-    required bool useWebView,
-    required bool enableJavaScript,
-    required bool enableDomStorage,
-    required bool universalLinksOnly,
-    required Map<String, String> headers,
-    String? webOnlyWindowName,
-  }) async {
-    this.url = url;
-    return true;
-  }
 }
